+++ doc/examples/effects/infinite_effects/lib/square.dart
@@ -3,14 +3,14 @@ import 'package:flame/components/position_component.dart';
 
 import 'dart:ui';
 
+import 'package:flame/extensions/vector2.dart';
+
 class Square extends PositionComponent {
   final Paint _paint;
 
-  Square(this._paint, double x, double y, {double angle = 0.0}) {
-    width = 100;
-    height = 100;
-    this.x = x;
-    this.y = y;
+  Square(this._paint, Vector2 position, {double angle = 0.0}) {
+    size = Vector2.all(100.0);
+    this.position = position;
     this.angle = angle;
     anchor = Anchor.center;
   }
@@ -18,6 +18,6 @@ class Square extends PositionComponent {
   @override
   void render(Canvas canvas) {
     super.render(canvas);
-    canvas.drawRect(toOriginRect(), _paint);
+    canvas.drawRect(size.toRect(), _paint);
   }
 }