+++ doc/examples/effects/sequence_effect/lib/main.dart
@@ -23,7 +23,7 @@ class MyGame extends BaseGame with TapDetector {
 
   MyGame() {
     final green = Paint()..color = const Color(0xAA338833);
-    greenSquare = Square(green, 100, 100);
+    greenSquare = Square(green, Vector2.all(100));
     add(greenSquare);
   }